pyproject.toml
[tool.poetry]
name = "blueprint"
version = "0.3.0"
description = ""
authors = ["Cody Jarrett <cody-jarrett@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
toml = "^0.10.2"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^3.0.0"
black = "^21.9b0"
pre-commit = "^2.15.0"
isort = "^5.9.3"
autoflake = "^1.4"
mkdocs-material = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Set isort profile to use black module
[tool.isort]
profile = "black"